+++ NEWS.md
@@ -1,5 +1,10 @@
 # cbbinom (development version)
 
+## Bug fixes
+
+* Fix a bug in `gen_hypergeo()` that may change input `U` and `L`.
+
+
 # cbbinom 0.1.0
 
 * Initial CRAN submission.